requirements.txt
loadenv
flask
Azure-Core
Azure-Search
Azure-Search-Documents